requirements.txt
django==2.1.2
psycopg2-binary==2.7.5
channels==2.1.5
djangorestframework==3.9.0
drf-yasg==1.11.0
drf-yasg[validation]
django-guardian==1.4.9
django-post_office==3.0.4
django-cors-headers==2.4.0
eventbrite==3.3.3
django-redis==4.9.0
channels_redis==2.3.1
sentry-sdk==0.5.1
django-filter==2.0.0